refactor(PlayerCards): card 객체에 대한 일급컬렉션 구현
